Dismemberment update - Partial Mapbase merge - Misc bug fixes

+++ src/game/server/baseentity.cpp
@@ -5525,7 +5525,7 @@ public:
 		{
 			const char *target = "", *action = "Use";
 			variant_t value;
-			int delay = 0;
+			float delay = 0;
 
 			target = STRING( AllocPooledString(command.Arg( 1 ) ) );
